aws-prod/requirements.txt
# flask
# flask-cors
# flask-session
# redis
# requests
# # grpcio
# # grpcio-tools
# scikit-learn
# pandas

# flask
# flask-cors
# flask-session
# redis
# requests
# scikit-learn
# pandas
# numpy
# kafka-python
# xgboost
# kaggle

flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
flask-session==0.4.0
redis==3.5.3
requests==2.26.0
scikit-learn==0.24.2
pandas==1.3.3
numpy==1.21.2
kafka-python==2.0.2
xgboost==1.4.2
kaggle==1.5.12
backoff==1.11.1